+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+and this project adheres to
+[Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
 
@@ -10,8 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- `Mail::new` now accepts a list of `to` addresses and thus supports sending an email to multiple
-  recipients.
+- `Mail::new` now accepts a list of `to` addresses and thus supports sending an
+  email to multiple recipients.
 
   ```rs
   // Before: